#1 stringformat bug: when stringformat is defined in multibinding, we can't do type cast variable to ex. string.

+++ WpfExample/ExampleViewModel.cs
@@ -10,8 +10,8 @@ namespace WpfExample
 {
     public class ExampleViewModel:object, INotifyPropertyChanged
     {
-        private float a = 10;
-        public float A
+        private double a = 10;
+        public double A
         {
             get { return a; }
             set
